Reordering code
