+++ MercadoPagoSDK/MercadoPagoSDK/Flows/OneTap/UI/PXOneTapViewModel.swift
@@ -357,8 +357,8 @@ extension PXOneTapViewModel {
 
         if let chargeRuleMessage = getChargeRuleBottomMessage(paymentTypeId), (status?.isUsable() ?? true) {
             let text = PXText(message: chargeRuleMessage, backgroundColor: nil, textColor: nil, weight: nil)
-            text.setDefaultTextColor(defaultTextColor)
-            text.setDefaultBackgroundColor(defaultBackgroundColor)
+            text.defaultTextColor = defaultTextColor
+            text.defaultBackgroundColor = defaultBackgroundColor
             return PXCardBottomMessage(text: text, fixed: false)
         }
 
@@ -372,8 +372,8 @@ extension PXOneTapViewModel {
 
         if reimbursementAppliedInstallments.contains(selectedInstallments), (status?.isUsable() ?? true) {
             let text = PXText(message: benefits?.reimbursement?.card?.message, backgroundColor: nil, textColor: nil, weight: nil)
-            text.setDefaultTextColor(defaultTextColor)
-            text.setDefaultBackgroundColor(defaultBackgroundColor)
+            text.defaultTextColor = defaultTextColor
+            text.defaultBackgroundColor = defaultBackgroundColor
             return PXCardBottomMessage(text: text, fixed: false)
         }
