+++ MercadoPagoSDK/MercadoPagoSDK/Models/PXText.swift
@@ -7,6 +7,7 @@
 
 import Foundation
 
+@objcMembers
 public class PXText: Codable, Equatable {
 
     let message: String?
@@ -23,7 +24,7 @@ public class PXText: Codable, Equatable {
         case weight
     }
 
-    init(message: String?, backgroundColor: String?, textColor: String?, weight: String?) {
+    public init(message: String?, backgroundColor: String?, textColor: String?, weight: String?) {
         self.message = message
         self.backgroundColor = backgroundColor
         self.textColor = textColor
@@ -34,13 +35,13 @@ public class PXText: Codable, Equatable {
         return lhs.message == rhs.message && lhs.backgroundColor == rhs.backgroundColor && lhs.textColor == rhs.textColor && lhs.weight == rhs.weight && lhs.defaultTextColor == rhs.defaultTextColor && lhs.defaultBackgroundColor == rhs.defaultBackgroundColor
     }
 
-    func setDefaultBackgroundColor(_ color: UIColor) {
-        self.defaultBackgroundColor = color
-    }
-
-    func setDefaultTextColor(_ color: UIColor) {
-        self.defaultTextColor = color
-    }
+//    func setDefaultBackgroundColor(_ color: UIColor) {
+//        self.defaultBackgroundColor = color
+//    }
+//
+//    func setDefaultTextColor(_ color: UIColor) {
+//        self.defaultTextColor = color
+//    }
 
     internal func getTextColor() -> UIColor {
         guard let color = self.textColor, color.isNotEmpty else {
